implement flag to load all scientific colormaps into paraview at once
Avoids misses due to fuzziness. e.g., when batlow and batlowW are both
initially not in ParaView, but only in ScientificColormaps, the first
one searched is loaded into ParaView. Because of the fuzziness, the next
search will end with some leniency on the first loaded colormap, without
looking to see if there's a perfect match outside of ParaView.

This hack just loads all the non-loaded presets into ParaView.